Run multiple DAEDALUS scenarios

Run multiple DAEDALUS scenarios

Usage

run_scenarios(
  country,
  infection,
  response_strategy = "none",
  response_time_start = 0,
  response_time_end = 0,
  duration = 100
)

Arguments

country

A country or territory object of class <daedalus_country>, or a country or territory name from those included in the package; see country_names, or a country ISO2 or ISO3 code; see country_codes_iso2c and country_codes_iso3c . Country-specific data such as the community and workplace contacts, the demography, and the distribution of the workforce into economic sectors is automatically accessed from package data for the relevant country name if it is passed as a string. To override package defaults for country characteristics, pass a <daedalus_country> object instead. See daedalus_country() for more.

infection

Similar to daedalus(), may be a character vector from among epidemic_names or a <daedalus_infection> object. May also be a list of <daedalus_infection> objects.

response_strategy

Either a string for the name of response strategy followed from among "none", "school_closures", "economic_closures", and "elimination", or a numeric vector of the same length as the number of economic sectors (45) giving the openness coefficient for each sector when closures are active.

response_time_start

A single number for the time at which the response_strategy comes into effect.

response_time_end

A single number of the time at which the response comes to an end.

duration

A vector of integer-ish numbers giving the durations over which to run scenarios. Each scenario is run for each duration.

Value

A <data.table>, potentially with data held in list-columns.